\name{pascals_triangle}
\alias{pascals_triangle}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
pascals_triangle
}
\description{
ky funksion printon trekendeshin e paskalit
}
\usage{
pascals_triangle(n)
\arguments{
  \item{n}{

}
}

\author{
Sabina Shkembi
}
